[add] add loader/function.rs
